Lifting State Up
State lifting is a pattern in React where you manage the state of a component in a higher-level or parent component. This allows you to share and pass down the state and related functions as props to child components. State lifting is commonly used when multiple components need access to the same piece of state or when a child component needs to update the state of a parent component.